+++ arni_gui/src/arni_gui/helper_functions.py
@@ -14,11 +14,11 @@ def choose_brush(index):
    
     # todo:can we assume this is always a string?
     #todo: add further keywords etc and check for errors. Is this the best possible way?
-    if index.data() == "Ok":
+    if index.data() == "ok":
         return QColor(127, 255, 0)
-    elif index.data() == "Warning":
+    elif index.data() == "warning":
         return QColor(255, 165, 0)
-    elif index.data() == "Error":
+    elif index.data() == "error":
         return QColor(255, 0, 0)
     else:
         return QColor(255, 255, 255)